+++ glibc/pthread/samples/pthread_create-deadline.c
@@ -13,7 +13,7 @@
 #include <sys/syscall.h>
 #include <pthread.h>
 
-#include "helpers.h"
+#include "pthread_helpers.h"
 
 #define gettid() syscall(__NR_gettid)